Force the runtest file to evaluate UInt64
The previous version would convert the result to Int64.

+++ exercises/grains/runtests.jl
@@ -4,8 +4,8 @@ include("grains.jl")
 
 @testset "On squares" begin
     @testset "On square $s" for s = UInt64(1):64
-        @test on_square(s) == 2^(s-1)
-        @test total_after(s) == 2^s - 1
+        @test on_square(s) == UInt64(2)^(s-1)
+        @test total_after(s) == UInt64(2)^s - 1
     end
 end
